Fixed issue #14704: DbStorage not found, API major version update Dev: can broke existing plugin Dev: move Storage to Storage directory
